split off jupyter into a separate module
