+++ jupyter/src/main/resources/META-INF/kotlin-jupyter-libraries/libraries.json
