+++ jupyter/src/main/resources/table.css
