+++ pom.xml
@@ -18,7 +18,7 @@
         <embedded-kafka.version>3.1.0</embedded-kafka.version>
         <spark3.version>3.2.1</spark3.version>
         <kotlin-jupyter-api.version>0.11.0-83</kotlin-jupyter-api.version>
-        <kotlinx.html.version>0.7.3</kotlinx.html.version>
+        <kotlinx.html.version>0.7.5</kotlinx.html.version>
         <hadoop.version>3.3.1</hadoop.version>
 
         <!-- Plugin versions -->
@@ -55,16 +55,6 @@
                 <groupId>org.jetbrains.kotlin</groupId>
                 <artifactId>kotlin-reflect</artifactId>
                 <version>${kotlin.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.jetbrains.kotlinx</groupId>
-                <artifactId>kotlin-jupyter-api</artifactId>
-                <version>${kotlin-jupyter-api.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.jetbrains.kotlinx</groupId>
-                <artifactId>kotlin-jupyter-test-kit</artifactId>
-                <version>${kotlin-jupyter-api.version}</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
